+++ testapp/src/main/java/com/rakuten/tech/mobile/testapp/ui/display/preload/PreloadMiniAppViewModel.kt
@@ -7,9 +7,9 @@ import com.rakuten.tech.mobile.miniapp.MiniAppSdkException
 import com.rakuten.tech.mobile.miniapp.permission.MiniAppCustomPermission
 import com.rakuten.tech.mobile.miniapp.permission.MiniAppCustomPermissionResult
 import com.rakuten.tech.mobile.miniapp.permission.MiniAppCustomPermissionType
-import com.rakuten.tech.mobile.testapp.ui.settings.AppSettings
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.launch
+import java.util.Locale
 
 class PreloadMiniAppViewModel(private val miniApp: MiniApp) : ViewModel() {
     private val _miniAppManifest = MutableLiveData<MiniAppManifest>()
@@ -22,7 +22,7 @@ class PreloadMiniAppViewModel(private val miniApp: MiniApp) : ViewModel() {
 
     fun checkMiniAppManifest(miniAppId: String, versionId: String) = viewModelScope.launch(Dispatchers.IO) {
         try {
-            val miniAppManifest = miniApp.getMiniAppManifest(miniAppId, versionId)
+            val miniAppManifest = miniApp.getMiniAppManifest(miniAppId, versionId, Locale.getDefault().language)
             val downloadedManifest = miniApp.getDownloadedManifest(miniAppId)
             if (downloadedManifest != null && isManifestEqual(miniAppManifest, downloadedManifest) &&
                 isAcceptedRequiredPermissions(miniAppId, miniAppManifest))